llamaparser-example/requirements.txt
llama-parse 
llama-index-embeddings-fastembed
llama-index
fastembed
llama-index-vector-stores-qdrant
qdrant_client
llama-index-embeddings-ollama
llama_index-llms-ollama
python-dotenv
llama-index-llms-groq